requirements.txt
bme280
# ST7735
pillow